# frontend/src/api/openapi.gen.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "meshmon Service API",
    "description": "Operator-facing HTTP API for the meshmon service. Agent-facing endpoints (Protobuf) are documented separately in `crates/protocol/proto/meshmon.proto`.",
    "contact": {
      "name": "meshmon contributors"
    },
    "license": {
      "name": "AGPL-3.0-or-later",
      "identifier": "AGPL-3.0-or-later"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/session": {
      "get": {
        "tags": [
          "session"
        ],
        "summary": "`GET /api/session` — return the session probe body.",
        "description": "Wired behind the `login_required!` layer in [`crate::http::router`],\nso unauthenticated requests are short-circuited with a bare 401\nbefore reaching this handler. The 401 response is documented here\nso the generated OpenAPI schema tells SPA clients exactly what to\nexpect.",
        "operationId": "session",
        "responses": {
          "200": {
            "description": "Active session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            }
          },
          "401": {
            "description": "No active session"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "SessionResponse": {
        "type": "object",
        "description": "Minimal session probe response.\n\nWrite-only on the server — we only ever construct and serialize this\ntype, never parse one back in — so only `Serialize` is derived.",
        "required": [
          "version",
          "username"
        ],
        "properties": {
          "username": {
            "type": "string",
            "description": "Signed-in username. The SPA hydrates its auth store from this\nso a hard-refresh tab still knows who's signed in."
          },
          "version": {
            "type": "string",
            "description": "`CARGO_PKG_VERSION` of the running service — surfaced so the SPA\ncan warn when the user's loaded bundle is older than the server."
          }
        }
      },
...
    }
  }
}